// src/api/types/PainReportResponse.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/


export type PainReportResponse = {
    /**
     * @type integer | undefined, int64
    */
    feedbackId?: number;
    /**
     * @type string | undefined, date-time
    */
    date?: string;
    /**
     * @type string | undefined
    */
    patientName?: string;
    /**
     * @type string | undefined
    */
    patientPhone?: string;
    /**
     * @type string | undefined
    */
    courseTitle?: string;
    /**
     * @type integer | undefined, int32
    */
    dayNumber?: number;
    /**
     * @type string | undefined
    */
    painDetails?: string;
};